requirements.txt
anyio==3.6.1
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
et-xmlfile==1.1.0
fastapi==0.85.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.23.4
openpyxl==3.0.10
orjson==3.8.0
pandas==1.5.0
passlib==1.7.4
psycopg2-binary==2.9.4
pydantic==1.10.2
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.5
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
sklearn==0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
xgboost==0.90
